refactor: move shared interfaces from types to interfaces directory
