+++ src/core/path-resolution/services/destination-resolver-factory.ts
@@ -1,6 +1,6 @@
 import { injectable } from 'tsyringe';
 import DestinationResolver from './destination-resolver';
-import { Factory } from '@shared/types/factory';
+import { Factory } from '@shared/interfaces/factory';
 
 type DestinationResolverProps = {
     source: string[];
