+++ src/shared/interfaces/factory.ts
