+++ src/shared/interfaces/pretty-error.ts
@@ -1,0 +1,3 @@
+export default interface PrettyError {
+    formatForDisplay(): string;
+}
