+++ src/shared/types/error.ts
@@ -1,9 +1,3 @@
-// Interfaces
-export interface PrettyError {
-    formatForDisplay(): string;
-}
-
-// Types
 export type NodeError = Error & {
     code: string;
 };
